rename: LinkBase클래스 위치 file에서 link로 변경
