+++ src/main/java/com/soda/request/entity/RequestLink.java
@@ -1,6 +1,6 @@
 package com.soda.request.entity;
 
-import com.soda.common.file.LinkBase;
+import com.soda.common.link.LinkBase;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
